refactor(mergeMap/mergeScan): unify logic
Creates a new `mergeInternals` function that can be used to create a merge-strategy operator, with extensiblity points for accommodating the nuances of each operator.

In theory, we could use this in `concatMap`, etc, directly.
